Add 'DirectDataBufferInt' supporting NIO based BufferedImages
- Contains convenient method to create an NIO backed BufferedImage
  'BufferedImage createBufferedImage(int width, int height, int numComponents, Point location)'

Standalone test TestTiledPrintingNIOImageSwingAWT 'prints', i.e. renders offscreen,
the Frame into
  - a traditional array-backed BufferedImage
  - an NIO-backed backed BufferedImage

+++ src/test/com/jogamp/opengl/test/junit/jogl/tile/TiledPrintingAWTBase.java
@@ -155,8 +155,8 @@ public abstract class TiledPrintingAWTBase extends UITestCase implements Printab
             final int frameHeight= printContainer.getHeight();
             final double scaleComp72;
             {
-                final int frameBorderW = frameInsets.left + frameInsets.right;
-                final int frameBorderH = frameInsets.top + frameInsets.bottom;
+                final double frameBorderW = frameInsets.left + frameInsets.right;
+                final double frameBorderH = frameInsets.top + frameInsets.bottom;
                 final double sx = pf.getImageableWidth() / ( frameWidth + frameBorderW ); 
                 final double sy = pf.getImageableHeight() / ( frameHeight + frameBorderH );
                 scaleComp72 = Math.min(sx, sy);
